+++ plugin/src/main/groovy/org/openbakery/TestFlightPluginExtension.groovy
@@ -1,0 +1,6 @@
+package org.openbakery
+
+
+class TestFlightPluginExtension {
+    def String outputDirectory = "build/testflight"
+}