+++ plugin/src/main/groovy/org/openbakery/XcodePlugin.groovy
@@ -16,9 +16,10 @@ class XcodePlugin implements Plugin<Project> {
 		project.extensions.create("keychain", KeychainPluginExtension)
 		project.extensions.create("provisioning", ProvisioningPluginExtension)
 		project.extensions.create("infoplist", InfoPlistExtension)
-		project.extensions.create("hockeykit", HockeyKitPluginExtension)
+        project.extensions.create("hockeykit", HockeyKitPluginExtension)
+        project.extensions.create("testflight", TestFlightPluginExtension)
 
-		Task keychainCreate = project.task('keychain-create', type: KeychainCreateTask)
+        Task keychainCreate = project.task('keychain-create', type: KeychainCreateTask)
 		Task xcodebuild = project.task('xcodebuild', type: XcodeBuildTask)
 		Task infoplistModify = project.task('infoplist-modify', type: InfoPlistModifyTask)
 		Task provisioningInstall = project.task('provisioning-install', type: ProvisioningInstallTask)
@@ -26,6 +27,7 @@ class XcodePlugin implements Plugin<Project> {
 		Task hockeyKitManifest = project.task("hockeykit-manifest", type: HockeyKitManifestTask)
 		Task hockeyKitArchiveTask = project.task("hockeykit-archive", type: HockeyKitArchiveTask)
 		Task hockeyKitImageTask = project.task("hockeykit-image", type: HockeyKitImageTask)
+        Task testFlightPrepare = project.task("testflight-prepare", type:TestFlightPrepareTask)
 
 		Task hockey = project.task("hockeykit")
 		hockey.description = "Creates a build that can be deployed on a hockeykit Server"
@@ -40,6 +42,7 @@ class XcodePlugin implements Plugin<Project> {
 		hockeyKitManifest.setGroup(GROUP_NAME)
 		hockeyKitArchiveTask.setGroup(GROUP_NAME)
 		hockeyKitImageTask.setGroup(GROUP_NAME)
+        testFlightPrepare.setGroup(GROUP_NAME)
 
 		hockeyKitArchiveTask.dependsOn(archive)
 		archive.dependsOn("clean")
@@ -49,11 +52,14 @@ class XcodePlugin implements Plugin<Project> {
 		Task xcodebuildCleanup = project.task('clean', type: XcodeBuildCleanTask)
 		Task provisioningCleanup = project.task('provisioning-clean', type: ProvisioningCleanupTask)
 		Task hockeyKitCleanTask = project.task('hockeykit-clean', type: HockeyKitCleanTask)
+        Task testFlightClean = project.task("testflight-clean", type: TestFlightCleanTask);
 
-		xcodebuildCleanup.dependsOn(keychainCleanup)
+        xcodebuildCleanup.dependsOn(keychainCleanup)
 		xcodebuildCleanup.dependsOn(provisioningCleanup)
 		xcodebuildCleanup.dependsOn(hockeyKitCleanTask)
+        xcodebuildCleanup.dependsOn(testFlightClean);
 		xcodebuildCleanup.setGroup(GROUP_NAME)
+
 
 		Task codesign = project.task('codesign', type: CodesignTask)
 		codesign.setGroup(GROUP_NAME)